style/refactor(client): Cleaned up code, move tab styling to a separate file
